core: cells: Move distance functions
